colpali_mcp_server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
elasticsearch
torch
transformers
accelerate
Pillow
colpali-engine @ git+https://github.com/illuin-tech/colpali.git
mcp
